requirements.txt
undetected-chromedriver 3.5.5
selenium                4.27.1
requests                2.32.3
Django                  5.1.4 
djangorestframework     3.15.2
beautifulsoup4          4.12.3   